agents: refactors testsys agents into a single crate
Consolidates all testsys agents into a single crate
