+++ agent/resource-agent/Cargo.toml
@@ -9,15 +9,15 @@ agent-common = { path = "../agent-common" }
 async-trait = "0.1"
 model = { path = "../../model" }
 log = "0.4"
-serde = { version = "1", features = [ "derive" ] }
+serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 snafu = "0.6"
-tokio = { version = "1", default-features = false, features = [ "time" ] }
+tokio = { version = "1", default-features = false, features = ["time"] }
 
 [dev-dependencies]
 env_logger = "0"
 nonzero_ext = "0.3"
-tokio = { version = "1", default-features = false, features = [ "macros", "rt-multi-thread" ] }
+tokio = { version = "1", default-features = false, features = ["macros", "rt-multi-thread"] }
 
 [build-dependencies]
 yamlgen = { path = "../../yamlgen" }
