--- agent/sonobuoy-test-agent/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "sonobuoy-test-agent"
-version = "0.1.0"
-edition = "2018"
-
-[dependencies]
-agent-common = { path = "../agent-common" }
-async-trait = "0.1"
-base64 = "0.13.0"
-log = "0.4"
-model = { path = "../../model" }
-test-agent = { path = "../test-agent" }
-tokio = { version = "1", default-features = false, features = [ "macros", "rt-multi-thread" ] }
-serde = { version = "1", features = [ "derive" ] }
-serde_json = "1"
-simplelog = "0.10.0"
-snafu = "0.6"
-
-[build-dependencies]
-yamlgen = { path = "../../yamlgen" }
