+++ agent/test-agent/Cargo.toml
@@ -9,13 +9,13 @@ agent-common = { path = "../agent-common" }
 async-trait = "0.1"
 model = { path = "../../model" }
 log = "0.4"
-serde = { version = "1", features = [ "derive" ] }
+serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 snafu = "0.6"
-tokio = { version = "1", default-features = false, features = [ "time" ] }
+tokio = { version = "1", default-features = false, features = ["time"] }
 
 [dev-dependencies]
-tokio = { version = "1", default-features = false, features = [ "macros", "process", "rt-multi-thread" ] }
+tokio = { version = "1", default-features = false, features = ["macros", "process", "rt-multi-thread"] }
 
 [build-dependencies]
 yamlgen = { path = "../../yamlgen" }
