+++ bottlerocket-agents/src/bin/ec2-resource-agent/ec2_provider.rs
@@ -4,7 +4,7 @@ use aws_sdk_ec2::model::{
     TagSpecification,
 };
 use aws_sdk_ec2::Region;
-use ec2_resource_agent::ClusterInfo;
+use bottlerocket_agents::ClusterInfo;
 use model::{Configuration, SecretName};
 use resource_agent::clients::InfoClient;
 use resource_agent::provider::{
