+++ bottlerocket-agents/src/bin/ec2-resource-agent/main.rs
@@ -1,6 +1,6 @@
-mod provider;
+mod ec2_provider;
 
-use crate::provider::{EksCreator, EksDestroyer};
+use crate::ec2_provider::{Ec2Creator, Ec2Destroyer};
 use resource_agent::clients::{DefaultAgentClient, DefaultInfoClient};
 use resource_agent::error::AgentResult;
 use resource_agent::{Agent, BootstrapData, Types};
@@ -15,7 +15,6 @@ async fn main() {
             std::process::exit(1);
         }
     };
-
     if let Err(e) = run(data).await {
         eprintln!("{}", e);
         std::process::exit(1);
@@ -27,7 +26,6 @@ async fn run(data: BootstrapData) -> AgentResult<()> {
         info_client: PhantomData::<DefaultInfoClient>::default(),
         agent_client: PhantomData::<DefaultAgentClient>::default(),
     };
-
-    let agent = Agent::new(types, data, EksCreator {}, EksDestroyer {}).await?;
+    let agent = Agent::new(types, data, Ec2Creator {}, Ec2Destroyer {}).await?;
     agent.run().await
 }
