+++ bottlerocket-agents/src/bin/eks-resource-agent/main.rs
@@ -1,6 +1,6 @@
-mod provider;
+mod eks_provider;
 
-use crate::provider::{Ec2Creator, Ec2Destroyer};
+use crate::eks_provider::{EksCreator, EksDestroyer};
 use resource_agent::clients::{DefaultAgentClient, DefaultInfoClient};
 use resource_agent::error::AgentResult;
 use resource_agent::{Agent, BootstrapData, Types};
@@ -15,6 +15,7 @@ async fn main() {
             std::process::exit(1);
         }
     };
+
     if let Err(e) = run(data).await {
         eprintln!("{}", e);
         std::process::exit(1);
@@ -26,6 +27,7 @@ async fn run(data: BootstrapData) -> AgentResult<()> {
         info_client: PhantomData::<DefaultInfoClient>::default(),
         agent_client: PhantomData::<DefaultAgentClient>::default(),
     };
-    let agent = Agent::new(types, data, Ec2Creator {}, Ec2Destroyer {}).await?;
+
+    let agent = Agent::new(types, data, EksCreator {}, EksDestroyer {}).await?;
     agent.run().await
 }
