+++ testsys/Cargo.toml
@@ -7,7 +7,7 @@ publish = false
 [dependencies]
 base64 = "0.13.0"
 env_logger = "0.9"
-kube = { version = "0.59.0", default-features = true, features = [ "derive"] }
+kube = { version = "0.59.0", default-features = true, features = ["derive"] }
 kube-runtime = "0.59.0"
 k8s-openapi = { version = "0.13.0", features = ["v1_20", "api"], default-features = false }
 log = "0.4"
@@ -22,7 +22,7 @@ tokio = { version = "1", features = ["macros", "rt-multi-thread"] }
 model = { path = "../model" }
 # regenerate testsys.yaml if the model has changed
 yamlgen = { path = "../yamlgen" }
-sonobuoy-test-agent = { path = "../agent/sonobuoy-test-agent" }
+bottlerocket-agents = { path = "../bottlerocket-agents" }
 
 [dev-dependencies]
 assert_cmd = "2.0"
